requirements.txt
astroid==2.2.5
certifi==2019.3.9
colorama==0.4.1
cycler==0.10.0
freetype-py==2.1.0.post1
isort==4.3.20
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.0
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
pep8==1.7.1
Pillow==6.0.0
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scipy==1.3.0
six==1.12.0
tornado==6.0.2
typed-ast==1.3.5
wincertstore==0.2
wrapt==1.11.1
